feat: JSON 파일 추가 및 추가된 JSON 파일불러오기 구현
item.json, items.json 파일 UnitTests 폴더에 추가
MockURLSession().dataTask 메서드 내부에서 json 파일 읽고 Data 타입으로
반환하는 기능 구현
